% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage-package.R
\docType{package}
\name{myPackage-package}
\alias{myPackage}
\alias{myPackage-package}
\title{myPackage: package juste pour voir comment on fait un package}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Claire Gayral \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
